requirements.txt
telethon==1.39.0
google-genai==1.7.0
Pillow==10.2.0
python-dotenv==1.0.1
docx2pdf==0.1.8
pandas==2.2.3
openpyxl==3.1.2
fpdf==1.7.2
python-pptx==1.0.2